Fix multi-ddata for aligned arrays
Previously multi-ddata's per-dimension-range (basically a range that represents
a sub-chunk) wasn't preserving the alignment of the original array. This caused
sporadic timeouts under numa for a program like:

   var SA: [1..3 by 2  align 2] string = ["align 2"];
   var SB: [1..3 by -2 align 2] string = ["reverse iter"];

We were only seeing this under numa and only for ~1/1000 runs, so I had been
worried that this was a startup race in our qthreads shim, or a bug in the
'distrib' schedulers mccoy threads. Luckily though, it was just a bogus free
manifesting itself strangely.

The problem was that we were ended up trying to call a destructor and free
memory that was never initialized. Under CHPL_MEM=jemalloc, this meant we were
sometimes trying to acquire a lock that was already held, leading to deadlock.

Under CHPL_MEM=cstdlib we got some hard failures, and by hacking fifo I was
able to run under valgrind and discover that we were trying to free
uninitialized memory.

The problem code comes from when we're trying to compute the size of each chunk
in order to free it. We do:

    chunkSize = numElts / mdRLen * mData(chunk).pdr.length;

and mData(chunk).pdr was just `1..3 by 2` instead of `1..3 by 2 align 2` so we
were computing the length as 2 instead of 1, which led to trying to call
chpl__autoDestroy on `dataChunk(0)[1]`, which in turn tried to destroy and free
memory that it shouldn't.

This just corrects the mulit-ddata code to maintain alignment.