Speed up reduction stress test by eliminating repetitive array building
reductions-stress*.chpl claims to be a reduction stress test, but almost all of
its runtime was spent building up an associative domain. This just moves the
building of the associative domain from the inner trial loop to the top level.

This dramatically reduces the runtime on chapcs:
    reductions-stress-mf   : 12.1s => 1.0s
    reductions-stress-fast : 15.5s => 0.1s

And we could probably bump the fast trial size and make it even more of a
reductions stresser now, but I'm not doing that here.