Vagrant testing scripts & Vagrantfiles
I've been using this setup to do portability testing for recent
releases across a variety of free unix distributions.

I recently got it automated enough that I'm willing to share it.
I hope it is useful to others.

I've written a README.txt outlining what this directory consists
of and how to use it.