See what happens if int(64)s can coerce to real(32)s
The compiler was explicitly preventing int(64)s from coercing to
real(32)s.  Should it?  I tried backing this out to see what, if
anything, breaks.  I think this was done because (a) we felt that ints
should coerce to _some_ floating point types, but we wanted to be
cognizant of the fact that there are int values that floating point
types can't represent accurately.  So we had them only coerce to
real(64)s in order to minimize the number of cases where loss would
occur without completely preventing them from coercing.

An intermediate point between what's on master today and opening the
gates up this wide would be to permit integers to param coerce to
real(32)s when they fit within the real(32)'s mantissa without loss of
accuracy (?).