Align atomics (#5915)
Align atomics (#5915)

[reviewed by @ronawho ]

64-bit align 64-bit atomic objects on 32-bit systems.

On some 32-bit systems, the max forced alignment is 32 bits. This means some 64-bit objects will be misaligned, and therefore some of those will straddle cache line boundaries. This is not a problem unless those objects are atomic. Then there is a serious performance penalty.

For the locks implementation of atomics, hardware atomic instructions are not used, so we do not need to worry about alignment. For the cstdlib implementation of atomics, the compiler vendors are supposed to choose the alignment optimally for each atomic type, relieving us of the need to do so. For the intrinsics implementation of atomics, we need to force 64-bit atomic objects to be at least 64-bit aligned, to guarantee that they will never straddle a cache line boundary on 32-bit systems. This change does that.

Implementation note: C11's alignment attribute will not work on typedefs, so we need the gcc alignment attribute, which is supported by all the compilers that use the intrinsics implementation of atomics.

Tested on old and new versions of: clang, Cray compiler, gcc, Intel compiler.
Tested clang and gcc on both 32-bit and 64-bit systems.