Fix a linux/darwin difference in DateTime module
The "%Y" (year) format specifier for function strftime zero-pads to 4 digits
on darwin, but doesn't zero-pad on Linux. Explicitly zero-pad it to 4 digits
in the isoformat function.

Removed two .notest files that were added due to this difference.

Changed a call to the non-existent gmtoff() into a call to utcoffset().