Make ReplicatedDist behave more traditionally
Rewrite operations so that rather than doing everything across
multiple locales, they operate on the local copy only.  The
only things which are now multilocale should be:

* allocating and deleting domains and arrays
* re-assigning the domain's index set

This permitted dramatic simplification to most of the arithmetic
robustness testing -- the removal of most replicated.good files,
retirement of .default.good files on said tests, and removal of
CLEANFILES, PREDIFF scripts, and .gitignore entries.