Skip man page testing for perf configurations
Yesterday, I widened the skipif for man page testing to support
testing in general configurations.  This broke for Cray performance
configurations pointing out that we really don't want/need to run man
page testing on a performance testing run (I often think that this may
be true for any test directories that contain a manual sub_test).

It seems likely that the man page testing would also fail on a Cray
correctness testing run, but that because we don't test the full
suite on Crays, we don't see the failure.  I'd guess that the failure
is likely due to a portability issue in the script or perhaps an
assumption that it's testing on a clone of the Git repo rather than
on the build module (?).