Fix a problem with removing unused functions at the end of resolution
When an unused function is removed from the AST at the end of resolution,
if it had a record defined in it, the record definition gets removed as
part of the body of the function. The autoCopy, autoDestroy and refType
for the removed type get left behind and continue to refer to the removed
type.

Before removing an unused function at the end of resolution, walk the
function's body looking for types defined within. If any are found, remove
the autoCopy, autoDestroy, and refType if they exist. Also remove the
autoCopy and autoDestroy functions for the refType.

Removed .future files for types/records/diten/unusedRecodInProc and
types/records/diten/unusedExternRecordInProc.