Add '(pre-release)' label to version output with SHA
For some time now, we've been discussing bumping our compiler
version number just after a release rather than just before it.
In doing so, we wanted to clearly label the version as being
a pre-release version.  This mod notes that this condition is
met whenever we print the SHA of the GitHub repo and simply
extends the logic that prints the SHA with a print of "(pre-release)".
It also updates the script that creates the .good file for our
test of the version number similarly.