Add new "fast" version of mandelbrot to release
This cleans up mandelbrot-blc.chpl and adds it to the release as a
'fast' (read: uglier but faster) variation of the benchmark.  It adds
mandelbrot-fast to the README and Makefile and to the two main
mandelbrot*.graph files.

While here, I also tried a style change to mandelbrot-blc.chpl
(zippering over the image domain and array to avoid some silly
manipulation in writing the image) to see how it affects performance.