Increment release version just after a release, not just before.
This change increments the release version numbers embedded in the
source, from 1.15.0 to 1.16.0, and adds "pre-release" markings to
some of those version numbers. Going forward, the "pre-release"
markings will be maintained on master branch, and removed on each
new release branch as it is created from master.

The goal is to make intermediate versions of Chapel built from GitHub
look more like a logical step between releases and not a version of
the last public release. We also want to clearly mark such versions
as being "pre-release" versions; i.e., not to be confused with the
official release to come.

This change also means that almost all documentation changes needed
in the next release can and should be made on master; reducing the
number of source changes that must be made on the release branch.