Attempt to make mandelbrot-fast testing more robust
There's a numerical instability in the CLBG benchmark that varies
somewhat unpredictably based on compiler, platform, etc.  In the past,
we've tried to call out such cases in skipif files, but this gets to
be difficult to maintain, particularly across compilers, processors,
etc.  This instability is discussed a bit here:

  https://alioth.debian.org/forum/forum.php?thread_id=14967&forum_id=2965&group_id=100815

and has plagued us in the past:

  https://github.com/chapel-lang/chapel/pull/5175

Here, we take a slightly different approach by committing both of the
.good files that we see and using a .prediff file to copy the right
one into the expected .good file's location.  The current logic is:

* if we're on the shootout box, use mandelbrot-fast-clbg.good so that
  we don't accidentally submit code that will break our official entry

* otherwise, speculatively compare against mandelbrot-fast-alt.good.
  If that passes, use that, otherwise use mandelbrot-fast-clbg.good.