Make array I/O more locality sensitive
Prior to this change, printing replicated arrays wouldn't work
as expected for the following reason:

* say we were printing from Locales[2]
* the generic I/O code would use an on-clause to go to the locale
  where the channel lives.  Say it's on Locales[0]
* we'd then print that replicand

Here, I use a new feature that Michael has prototyped in #5989
in which the originating locale for a write()/read() can be
passed to the readThis()/writeThis() routine.  Specifically,
I expand the ChapelArray read/write routines to accept this
locale argument and then to pass it along to the dsi routines
if they want it (as determined by reflection).

The other challenge here was to get array views working properly
for replicated distributions.  It turned out that I'd axed
some code more than I should've in previous revisions on this
branch.  Here, I restore some of that to make sure the originating
locale passes through the routines.  This required adding a locale
argument to my default rectangular utility write routines.