Broaden special-casing of mandelbrot.chpl to handle mandelbrot*.chpl
This is a silly test which seems to be designed to dodge around the
fact that mandelbrot's output is binary and therefore cannot easily be
grepped.  Previously, we checked for mandelbrot.chpl.  This change
makes us check for mandelbrot*.chpl.  A better test might be to see if
outfile is binary, but unfortunately I can't afford to take more test
failures this week so am leaving a comment instead.