Remove release version number from GOALS
There has never been a change to the actual public release goals
set forth in this file. (The changes have been mechanical details
like file formatting or pointers to external references- and
incrementing the release version number, twice a year).
That being the case, there seems little reason to maintain the
release version number in this file.
If the goals do change in a meaningful way, then it would make
sense to add the effective Chapel release version for the new
goals.