Fine-tune Elliot's chameneos
Something about last night's changes (the storage of 'color' as a field
of 'Population'?) made Elliot's chameneos perform much more poorly than
it had been.  Here, I remove the color field, replacing it with a
constructor (I tried to use an initializer but ran afoul of issue #5935).
The constructor-based approach seemed to require the declaration of a
domain field as well -- i.e., I couldn't find a way to simply make
'chameneos' generic and have the constructor resolve.  I also changed
the deinitializer to a destructor for symmtry.

This version seems to be going faster than it has before, and faster
than my variation as well.  Unfortunately, it relies on
constructors/destructors which will be going away soon, so will need
to be updated in the future.