File an idealized version of chameneos as a future
This is a variation on chameneos-ejr which attempts to use an
initializer rather than a constructor.  It currently runs afoul of
issue #5935, though I expect that once we get past that, there will be
other challenges related to generics ahead (e.g., am I now writing
cases that I told Mike I didn't expect us to be able to handle a few
weeks ago? :)  And what options would we have for writing this idiom
if not this way?)