Futures for 'Owned' types
I tried to write a version of Chameneos that used 'Owned' types today
and ran into a few failures.  The owned-init* cases look like true
bugs to me.  The owned-comp case may be a limit to what one can/cannot
be expected to be able to do with 'Owned' (i.e., maybe this is a
'feature request'), but it feels a little unfortunate...