Promote latest bradc study version to release version as 'fast' version
The current 'elliot' and 'bradc' variations of chameneos redux are
performing similarly, but my version doesn't rely on the use of a
slightly crufty record constructor (and initializers can't handle this
case yet).  For this reason, Elliot and I were thinking on Friday
that this is probably the version to put forward for the time being
and to consider switching to the Elliot version once it is supported
by initializers.