Functions that don't return values vs. functions that return '_void'
Since 'void' became a first class type, functions that do not return a value
have been treated the same as functions that return the value '_void'.  They
have been allowed to be used as variable initializers (making the variable
type 'void'), assigned to 'void' variables, and passed into generic
functions (making the argument type 'void').

Now, functions that don't have any return statement and functions that have
a return statement without arguments are treated as not returning a usable
value, while functions that explicitly return a '_void' value have a value
that can be used.

Changed the parser to allow 0-argument return statements instead of
translating them to return '_void'. Changed checkParsed to consider "return;"
to be equivalent to "return _void;" for some error checking purposes.

Added a new cleanup to the cleanup pass to translate 0-argument return statements to return '_void', but also to mark the function with a new flag
FLAG_NO_RETURN_VALUE.

Made normalize add the NO_RETURN_VALUE flag when it adds "return _void" to
functions that had no return statements.

Finally, in functionResolution, make resolveMove issue an error for moves
where the RHS is a function call that returns '_void' but has the
NO_RETURN_VALUE flag.