Revert .good files of tests that use results of functions that don't return
When first class void types were introduced, the tests changed here had
their .good files updated to reflect that a function that does not return
was considered the same as a function that explicitly returns '_void'.

Now that functions that don't return values are treated as different from
functions that return _void, revert back to the old error messages about
"illegal use of function that does not return a value".