Ignore nested functions when verifying initializers don't return values
Prior to this change, we weren't distinguishing between the initializer and any
functions that were defined within its bounds (whether compiler inserted or
otherwise).  Since under the covers we define conditional expressions as nested
functions, this meant that we were complaining about returning from those
functions when the initializer itself was not trying to return anything.

The fix for this was very simple: when gathering the call expressions to check
for return statements, only gather those whose parent symbol was the initializer
itself.  With this change, the future
test/classes/initializers/initViaConditional.chpl now passes.