Add more tests of phase 2 initializers with generics and update a .good
Last commit, I added a test of setting a param field during Phase 2 when the
param had a default value.  I had a mistake in my expected .good file, and the
error message that it generates today is not the one I want, so I added a .bad
and .future file for the test.  I also added two new tests: one of setting a
param field during Phase 2 when the param had a default type, and one of setting
a type field during Phase 2 when the type had a default value.  Again, both of
these cases error, but not in the way I would like.