Special handling for void returns in extern functions with array arguments
Extern functions that have array arguments have a wrapper function that
handles pulling the right fields out of the array type to pass to the
extern function. The wrapper returns the result of calling the extern
function. The wrapper is created early enough that an explicit return
type on the extern function is not yet resolved and available.

When this wrapper function is created to handle array arguments and the
the extern function returns void, treat the extern function as though it
returned the value '_void' so that the wrapper can use/return the result.
Treat the wrapper as a 'void' function that did not return a value so that
calls to it cannot use the returned void value.