Remove the mdBlk multi-ddata field for arrays.
mdBlk is the ratio between the strides at the mdParDim'th dimension for
a reindexed alias and an original array.  We need it so we can compute
block and offset-within-block for multi-ddata using the reindexed
alias's indices when the blocking is in terms of the original's indices.
But as of the array views work this ratio is no longer needed for arrays
themselves; it's only needed for array views.  Here, remove it from the
array descriptor, and also remove it from array address calculations.

This makes 5-10% improvements in several of the 'LCALS (raw, long)'
performance results for linux64.