Add general version of initCallsDom future
In looking at the bug behind initCallsDom, I had realized that the problem
wasn't specific to arrays at all - it was due to the compiler assuming the
generic argument for that initializer was allowed to be of the type on which
the initializer was defined, making it a candidate copy initializer.  Then,
when it tried some autoCopy/initCopy operations with that assumption, it ran
into issues because the method called on the argument was not defined for the
type it was trying to use.  However, copies aren't actually made of the type,
so trying to resolve the initCopy/autoCopy functions that make this bad call
is incorrect.  We've avoided this for other types that obviously didn't define
a copy initializer, but that avoidance was thwarted by having the argument
list appear like what we were expecting (by being generic).