Look at the where clause work around on the general case
When I was investigating initCallsDom, I tried using a where clause on the
initializer to avoid triggering the copy initializer attempt.  This ran into
another issue, which seems to be that there is a problem with certain storage
of array fields.  I realized today that I hadn't tried the where clause
work-around on the general case; when I did, I discovered that I had messed up
in creating the .good file for the general case future - I had written it when
I thought I was going to use a class and neglected to update it when I realized
that I needed to use records.

This commit fixes the expected output for the future, and adds the where clause
version of it (which passes)