Address valgrind failures from PR #5949
PR #5949 attempted to clean up the loop computing the maximum number of virtual
method table slots for any type.  That clean-up was causing valgrind errors
because the Virtual Method Table can contain keys that are no longer in the AST
(and in fact are pointers to freed memory). Therefore, switch to using a
pattern of traversing types that are in the AST and then 'get'ing the
appropriate VMT elements.