Allow iterators that explicitly yield '_void'
Change yield statements that yield values with 'void' type to yield the global
'_void' value.  Change for loop index variables that are type 'void' to refer
to the global '_void' value.

This gets two of the three cases in issue #5697 working as described there.

Add a test with iterators that explicitly yield _void. Add a .future for an
iterator that doesn't yield anything.