Use at least C++11 when recent enough compiler available (#6113)
Use at least C++11 when recent enough compiler available (#6113)

[Note:  There was post-meeting discussion where the original uneasiness was resolved.
Reviewed by @ronawho and @benharsh who had expressed reservations earlier.]

This does for the Chapel compiler what was already being done for the runtime. If we are using a C/C++ compiler recent enough to default to C11, then we ensure that we use at least C++11 to match. However, we do not demote to C++11 if the compiler defaults to something later.

This allows us to use some extra error checking, because fewer implicit type conversions are allowed for C++11 and beyond. It does not prevent us from testing with older standards; that only requires using an older compiler.

This shifts the amount of time to catch errors around a bit. Errors caught by C++11 will now be caught immediately, because clang has these checks. (G++ 7 will also perform the checking.) Errors caused by using C++11 features in the compiler will now be caught during the next nightly test instead of immediately.

This change is applied everywhere that we have this variable language standard scheme for the runtime: the clang, gnu, and intel compilers.

Performed build and paratest with CHPL_HOST_COMPILER set to clang, gnu, and intel.
Tested with gcc 4.9, 5.4, and 6.3, which fall into the three separate prongs of the Makefile logic.
Tested with the latest Apple clang (approx. clang 3.9/4.0) and clang 3.3, which exercise the only two cases that clang exhibits to date.
Tested with v17 of the intel compiler, which exercises the only case that intel exhibits to date.
Also tested with and without CHPL_LLVM=llvm.