Improve tests I previously defined
Add .future files for all the declaration tests I defined previously, so that
they don't result in testing failures (because I don't expect them to work yet).
Add .good files for the expected output of the no-arg-*.chpl files, because I
hadn't added any before.  Fix matches-var to not be uncertain about how to
define the type specifier for the var case.