Convert multi-ddata from a ddata-of-ddatas to a tuple-of-ddatas.
Instead of a ddata of chunks of array ddata, use a tuple of those
chunks.  This is like what was done for the remote access data a few
months ago.  An issue is that the tuple wants to use 1-based indexing,
while the chunking arithmetic wants to use 0-based indexing.  To get
around that for now, add an accessor function (mDataAcc()) that accepts
a 0-based chunk index and returns a 1-based chunk (mData element).

WHile here, add the same accessor function for the remote access data
and recode to use it.

Interestingly, this all slows down at least prk stencil serial by about
30%.