First pass at building an internal type function for generic initializers
No idea if it will work at all yet, but it's a good starting point.  From the
initializer, creates a type function with only an arg per generic argument in
the initializer.  In the body, has a call to the initializer we are building
this around, and returns the type of the instance initialized by that function.