Add a run_command variant that yields output as it runs
utils.run_command runs the command, and then returns the output. For long
running commands (like an upcoming paratest.chapcs rewrite), we need to be able
to output as we go. This adds a utils.run_live_command that does this. Note
that it merges stderr with stdout and it doesn't support stdin.

Based on a few stackoverflow posts:
 - http://stackoverflow.com/questions/4417546/constantly-print-subprocess-output-while-process-is-running/4417735
 - http://stackoverflow.com/questions/5582933/need-to-avoid-subprocess-deadlock-without-communicate