Fix the declaration test failures in numa
The output for potential matching functions was different for numa, but the
functions causing the problem weren't the important part of the output.  When
I take the chance to clean up our resolution failure output for methods, the
output for these tests will be consistent, but until then just sanitize it and
be done.