Add unique comm ID to runtime function declarations
A new argument, int32_t commID, has been added to the common get/put
functions (and their nonblocking/strided counterparts). It is before the
line and file numbers at gbt's suggestion.

The CHPL_UNKNOWN_COMM_ID macro is added to represent a commID whose
source is unknown.

Also addes the 'COMMID( i)' macro, which is basically identicall to
INT64. I chose this name to make it easier to search through the
generated code.