Update chpl-visual-debug.c with commID data
This commit increments the ChplVdebug version from 1.2 to 1.3.

The CHPL_VDEBUG_GETPUT_FORMAT_STRING macro is introduced to provide a
common format string for the get/put events. All of the get/put
functions were passing '1' to the elemSize slot, so I just made that '1'
character part of the format string for now. Also, according to
tools/chplvis/TextDataFormat.txt the typeIndex and length fields were in
the wrong order, so I swapped them.