Minor clean ups, that allow basic cases to work!
- I'd forgotten that the variable holding the result of the type specifier call
needed to be flagged as a type variable.
- Having the type specifier call "init" directly was problematic, so just have
it use a "new" instead
- With using the "new", I needed to not explicitly set the type of the local
variable in the type specifier, or I would get a type mismatch in the
assignment.

With this change, at least
test/classes/initializers/generics/declarations/matches-type.chpl now passes!
I still need to handle the case of multiple initializers leading to the same
type specifier, get the record implementation closer to working, etc.  But it's
a start!