Don't line buffer output in the paratest.chapcs python rewrite
We were previously doing `process.stdout.readline`, but for paratest this meant
the trailing smiley faces would be buffered until all testing was finished,
which gave the impression that paratest stopped working. Just read a character
at a time instead of a line at a time. This probably has terrible performance
for processes that have a huge amount of output, but paratest doesn't actually
output that much, so it's probably not a big deal.