Suppress I-fail-without-re2 test if using PGI
Current version of re2 requires C++11 atomic support, which PGI
does not have. So, with PGI, re2 does not get built and the
I-fail-without-re2 test fails.

This change adds I-fail-without-re2.suppressif to quiet this error.

PGI is only used in Cray Whitebox tests, so those are the only tests
that should be affected by this change.