Add variable to channel to store locale on which I/O started
I think this is a better alternative than adding a locale argument to
readThis/writeThis etc.

This commit adds a field _readWriteThisFromLocale to the channel and the
user-facing readWriteThisFromLocale to query it. It adds the internal
getIoStartedOnLocale helper method and calls that when figuring out the
original locale of some I/O operation.