Add new tests of the --library flag when using main()
Following up on Lydia's work last week, create a test that defines a
"library" and also includes a main().  I don't think that creating a
library from Chapel should preclude the ability to have a function
named main()... the function should still exist, but simply not serve
as the program's entry point.  For example, this might be useful for
making the library be self-testing when compiled without the --library
flag.