Quiet our use of the 'ar' command
When compiling in --library mode and generating a static .a library,
the 'ar' utility prints to the console the name of the library that
it's creating, which feels noisy because it's in the /tmp directory
with a temporary name (the compiler then moves it back to the output
directory, renaming it).  Adding the -c flag to quiet this message.