Change the error for defining main() in library mode into a warning
I don't think that it should be an error to have a function named
main() in a Chapel program when creating a library from it.  So
instead, change it to a warning indicating to the user that there's
nothing special about this procedure named main() (i.e., it's not the
program's entry point).