Fix sys/chpl alloc mismatch in runtime/src/qio/auxFilesys/hdfs/qio_plugin_hdfs.c
get_locale_name() was using the system strndup to allocate a string, and then
later freed it with qio_free (which maps to chpl_mem_free.) You can't free
memory allocated from a different allocator, so this is super wrong under
CHPL_MEM=jemalloc. The hdfs code isn't tested nightly, so it was probably only
exercised when the code was being worked on a few years ago when cstdlib was
our default allocator, in which case the mismatch could have easily gone
unnoticed.

This just uses qio_malloc+qio_memcpy instead of strndup. Also note that strndup
isn't portable, so this is also a portability improvement.