Get make install + --llvm  working
The --llvm compiles use the compileline script, so install that.
(It also might be useful in some cases when bulding C code to
 be integrated with Chapel).

Then, modify compileline to set CHPL_MAKE_RUNTIME_LIB,
CHPL_MAKE_RUNTIME_INCL, and CHPL_MAKE_THIRD_PARTY. It was
already setting CHPL_MAKE_HOME but these others need to be
set too in order for the installed Make to work correctly.