Replace qio_valloc with qio_memalign
qio_valloc was only used in one place, so provide memalign and align to
sys_page_size() at the callsite. Note that this is actually a change in
behavior. qio_valloc used to map down to chpl_valloc, which actually aligned to
the heap page size, not the system. I'm pretty sure this code wants to be
system page aligned, not heap page aligned (the allocation is way, way smaller
than heap pagesize), but I need to verify that with Michael.