More stderr redirects in start_test
Following the lead of #6183, I added a redirection of stderr to the log file
in the remaining cases where start_test uses subprocess.Popen().

Addendum to #6183: redirection of sub_test's stderr into the log file
(that it restored) was lost in #2185:

```
-   $sub_test "$compiler" |& $tee -a $logfile
+   p = subprocess.Popen([sub_test, compiler], stdout=subprocess.PIPE)
```

While there, this PR also:

* factors out subprocess.Popen machinery into run_and_log()
* ... which is now invoked by all Popen uses in the script
* renames run() to run_sub_test(), for clarity and better grepping
* tweaks an error message for clarity

Tested by running in a directory where I created
a stderr-producing and error-returning `sub_test`.