Move runtime privatization support into chpl-privatization.h
Move runtime privatization support from the .c file to the header.
chpl_getPrivatizedClass() can be called frequently, so we want to allow the
backend compiler to fully optimize/inline calls to it.

Moving the privatization source code into the header has a pretty big
performance impact for the stencil PRK, improving performance by about 15% for
16-node-xc. There's also some minor improvements for fft, and lulesh.

This is motivated by https://github.com/chapel-lang/chapel/issues/6184, though
it's not quite enough to close that issue yet.