Fix issue with accidentally creating the type again for new calls
Because type constructors used the old method of type instantiation, we weren't
logging the instantiation and so when "new Foo()" calls occurred we would
recreate the instantiation instead of reusing the one created by the type
constructor.  This change makes it so that we utilize the code path which will
log these types as we create them, to allow these types to be reused between
the two instantiation strategies.  With this change, all the tests I added in
the last commit pass.

I modified instantiateTypeForTypeConstructor slightly to take in an
AggregateType instead of a Type, since it relies on having that instantiation
anyways and I wanted to make my support method on AggregateTypes.